rest_api_server/requirements.txt
Django==3.1.6
djangorestframework==3.12.2
PyYAML==5.4.1
kubernetes==12.0.1
urllib3==1.26.3